requirements.txt
flask
scikit-learn
plotly
pylint
psx-data-reader
pandas
requests
datetime
